workshop-dlt/my-dlt-pipeline/pyproject.toml
[project]
name = "my-dlt-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dlt[workspace]>=1.22.0",
    "duckdb>=1.4.4",
    "ibis-framework>=12.0.0",
    "marimo>=0.19.11",
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
]
